requirements.txt
streamlit>=1.39.0
google-genai>=0.8.0
Pillow>=10.0.0
supabase>=2.9.0
